gpio control & fix bugs

+++ killAll.sh
@@ -1,8 +1,15 @@
 #!/bin/sh
 
-while read line
-do
-   kill $line
-done < pid.log
+file="pid.log"
 
-rm pid.log
+if [ -f $file ]
+then
+ while read line
+ do
+  kill $line
+ done < $file
+
+ rm pid.log
+else
+ echo "not already started"
+fi
